+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - `mj-carousel`: `thumbnails="supported"`
   - inherit `font-family` from `mj-accordion-element`
   - `width="auto"` for `mj-table`
+  - nonzero `cellspacing` for `mj-table`
 - support `mj-include` with `type="css"` and `type="html"`
 - support `mj-include` of files which contain an `<mj-head>`
 - support `<mj-class>` with child elements (default attributes for nested elements)